Solution Chapter 5 Task 21 (исправлены недочёты, добавлена проверка на пересечение шаром рёбер параллелепипеда)

+++ src/main/java/com/walking/intensive/chapter5/task21/Sphere.java
@@ -6,7 +6,7 @@ public class Sphere {
     final private double zAxisCoordinate;
     final private double radius;
 
-    public Sphere(int xAxisCoordinate, int yAxisCoordinate, int zAxisCoordinate, int radius) {
+    public Sphere(double xAxisCoordinate, double yAxisCoordinate, double zAxisCoordinate, double radius) {
         this.xAxisCoordinate = xAxisCoordinate;
         this.yAxisCoordinate = yAxisCoordinate;
         this.zAxisCoordinate = zAxisCoordinate;
